Refactor nav.module.scss
